+++ CHANGELOG.md
@@ -23,6 +23,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed an issue that would cause compilation errors if a Unity project using Yarn Spinner also used a DLL with the same name as one of Yarn Spinner's dependencies (for example Google Protocol Buffers).
   - The dependency DLLs that come with Yarn Spinner (for example, `Antlr.Runtime`, `Google.Probuf`, and others) have been renamed to have the prefix `Yarn.`, and the assembly definition files for Yarn Spinner have been updated to use the renamed files. 
   - Huge thanks to [@Sygan](https://github.com/Sygan) for finding and describing [the fix for this problem](https://github.com/YarnSpinnerTool/YarnSpinner-Unity/issues/15#issuecomment-1036162152)!
+- The `YarnProject.GetProgram()` method has been replaced with a property, `Program`.
+  - `GetProgram()` still exists, but has been marked as obsolete and will be removed in a future release of Yarn Spinner.
+  - `YarnSpinner.Program` has better performance, because it caches the result of de-serializing the compiled Yarn Program.
 
 ## [2.2.1] 2022-06-14